requirements.txt
langgraph
langchain-core
langchain-groq
requests
python-dotenv
pydantic
rich
mcp
fastmcp
httpx
asyncio
langchain_mcp
aiohttp
langchain-mcp-adapters